Send email notification with pb log and status
This commit adds the send email (with playbook log and exit status)
feature for those users that are not running the cronjob and want a mail
notification.

A new handler has been added and it is only executed when the
`atom_replication_email_notification_enabled` variable is enabled. This
handler is using the mail ansible module, and all its fields can be
defined (to, cc, body, subject, mailserver) except the attached file
that is configured with the ansible log file.

This handler will be run always because `force_handlers = True` is used
in `ansible.cfg`.

A new internal variable (`_playbook_status`) is defined at the beginning
of the role as "ERROR" and at the end of the role, if all goes fine, it
is redefined as "SUCCESS". This variable is sent as part of the email's
subject.